+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/notification/NotificationHandlerTest.kt
@@ -54,7 +54,7 @@ class NotificationHandlerTest {
             multiSettings,
             matrixClients,
             { callbackParam = it }
-        ) { _, _, _ ->
+        ) { _, _, _, _ ->
             notificationHandler
         }
     }